Nuevos Test: TestRetiroCantidadCeroNoPermitido, TestSaldoConIntereses

+++ TestProject/UnitTestInteres.cs
@@ -8,8 +8,8 @@ namespace TestProject
         [Fact]
         public void TestInteres1()
         {
-            decimal Monto = 100;
-            decimal Tasa = 0.2M;
+            double Monto = 100;
+            double Tasa = 0.2;
             double Resultado = Interes.CalculoInteres(Monto, Tasa);
             Assert.Equal(0.0546, Resultado, 4);
         }
@@ -18,8 +18,8 @@ namespace TestProject
         public void TestInteres2()
         {
             // Cuando es 0
-            decimal Monto = 0;
-            decimal Tasa = 0.2M;
+            double Monto = 0;
+            double Tasa = 0.2;
             double Resultado = Interes.CalculoInteres(Monto, Tasa);
             Assert.Equal(0, Resultado, 4);
         }
@@ -28,8 +28,8 @@ namespace TestProject
         public void TestInteres3()
         {
             // Cuando el valor es negativo
-            decimal Monto = -100;
-            decimal Tasa = 0.2M;
+            double Monto = -100;
+            double Tasa = 0.2;
             double Resultado = Interes.CalculoInteres(Monto, Tasa);
             Assert.Equal(0, Resultado, 4);
         }
